Extract mediator and view for view mode
